Implement CompositeDisposable that is a disposable container that can hold onto multiple other disposables
